Move expressions into their own file
